Fix WORKDIR and change ADD to COPY
Using WORKDIR automatically creates the directory and moves to that
directory. This means all COPY/ADDs will be relative to that.
Also, it is Docker best practice to use COPY instead of ADD because COPY
is more transparent, as it does not have some additional behavior that
ADD has. For more information, see
https://docs.docker.com/develop/develop-images/dockerfile_best-practices/#add-or-copy